Fix links to master, point to older version
This documentation was mostly written in 2016, and it uses a lot of the links to master, which are broken now. I replaced 'master' almost everywhere (all except the lab exercise link) with a now outdated version, such that the links and descriptions match as I read this. Feel free to pull (no conditions apply, no credit needed).